requirements.txt
Pyrogram
PyYAML
Telethon
jikanpy
aiohttp
requests
tgcrypto
